+++ src/Endpoints/Certificates.php
@@ -118,12 +118,13 @@ class Certificates extends Endpoint
      */
     public function update(Certificate $certificate): Response
     {
-        $this->validateRequired($certificate, 'update', ['id', 'cluster_id']);
+        $this->validateRequired($certificate, 'update', ['id', 'cluster_id', 'main_common_name']);
 
         $request = (new Request())
             ->setMethod(Request::METHOD_PATCH)
             ->setUrl(sprintf('certificates/%d', $certificate->getId()))
             ->setBody($this->filterFields($certificate->toArray(), [
+                'main_common_name',
                 'common_names',
                 'certificate',
                 'ca_chain',