+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 cluster API. See the changelog of the [cluster API](https://cluster-api.cyberfusion.nl/redoc#section/Changelog) for 
 detailed information.
 
+## [1.25.0]
+
+### Added
+
+- Add `bubblewrap_toolkit_enabled` to cluster.
+- Add `main_common_name` to certificate.
+
 ## [1.24.2]
 
 ### Fixed